+++ lib/language_pack/shell_helpers.rb
@@ -48,7 +48,7 @@ module LanguagePack
     end
 
     def self.blacklist?(key)
-      %w(PATH GEM_PATH GEM_HOME GIT_DIR JRUBY_OPTS JAVA_OPTS JAVA_TOOL_OPTIONS).include?(key)
+      %w(PATH GEM_PATH GEM_HOME GIT_DIR JRUBY_OPTS JAVA_OPTS JAVA_TOOL_OPTIONS RUBYOPT).include?(key)
     end
 
     def self.initialize_env(path)
@@ -177,7 +177,7 @@ module LanguagePack
       def output
         raise "no file name given" if @file.nil?
         exec_once
-        @file.read
+        @file.read.encode('UTF-8', invalid: :replace)
       end
 
       def timeout?